Add ItemTouch interfaces Move some files around
