+++ app/src/main/java/com/blackrose9/myjournal/adapter/CategoryAdapter.java
@@ -1,4 +1,4 @@
-package com.blackrose9.myjournal;
+package com.blackrose9.myjournal.adapter;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -6,6 +6,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.TextView;
+
+import com.blackrose9.myjournal.R;
 
 public class CategoryAdapter extends BaseAdapter {
     private Context mContext;
