+++ app/src/main/java/com/blackrose9/myjournal/util/ItemTouchHelperAdapter.java
@@ -1,0 +1,5 @@
+package com.blackrose9.myjournal.util;
+
+public interface ItemTouchHelperAdapter {
+    void onItemDismiss(int position);
+}
